pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0", "hatch-vcs>=0.4"]
build-backend = "hatchling.build"

[project]
name = "pwnkit"
dynamic = ["version"]
description = "Axura's reusable pwn utilities, gadgets, debugging, shellcodes, templates"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Axura", email = "asuraulord@gmail.com" }]
keywords = ["pwn", "ctf", "exploit", "heap", "rop", "pwntools", "template", "stack"]
requires-python = ">=3.8"
dependencies = [
  "pwntools>=4.11.0",
  "typing-extensions>=4.10.0; python_version<'3.11'",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Security",
  "Topic :: Software Development :: Libraries",
  "Environment :: Console",
]

[project.urls]
Homepage  = "https://4xura.com"
Repository = "https://github.com/4xura/pwnkit"
Issues     = "https://github.com/4xura/pwnkit/issues"
Changelog  = "https://github.com/4xura/pwnkit/releases"

[project.scripts]
pwnkit = "pwnkit.__main__:cli"

[project.optional-dependencies]
dev = ["pytest"]

# ---- Hatch (VCS version) ----
[tool.hatch.version]
source = "vcs"

# Write src/pwnkit/_version.py at build time
[tool.hatch.build.hooks.vcs]
version-file = "src/pwnkit/_version.py"

# src-layout
[tool.hatch.build.targets.wheel]
packages = ["src/pwnkit"]
# If you want to ensure templates land in the wheel:
# include = ["src/pwnkit/templates/*.tpl", "src/pwnkit/templates/*.py.tpl"]

[tool.hatch.build.targets.sdist]
include = [
  "src/pwnkit/**",
  "README.md",
  "LICENSE",
  "pyproject.toml",
  "tests/**",
]

[tool.hatch.build]
exclude = ["*.bak", "*.swp", ".venv", "dist", "build", ".pytest_cache", ".ruff_cache"]

# ---- Tooling ----
[tool.ruff]
line-length = 100
target-version = "py38"
extend-select = ["E", "F", "I", "UP"]
exclude = ["dist", ".venv"]

# ---- Pytest ----
[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
  "ignore:invalid escape sequence:SyntaxWarning:ropgadget\\.",
  "ignore:invalid escape sequence:SyntaxWarning:ropgadget\\..*",
]

# These setuptools sections are ignored by Hatch, but harmless;
[tool.setuptools.package-data]
pwnkit = ["templates/*.tpl", "templates/*.py.tpl"]

[tool.setuptools.packages.find]
where = ["src"]

